move sort function out of enterprise
