+++ client/web/src/insights/utils/sort-series-by-name.test.ts
